feat: general post fetcher endpoint
* remove duplicate code from endpoints

* posts.js endpoint takes the ids and returns the post data
